chore: move tests to ITJdbcTest
